pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kindling"
version = "0.1.0"
description = "A lightweight, profile-driven FHIR synthetic data generator"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Matt Berg", email = "matt@example.com"}
]
keywords = ["fhir", "synthetic", "data", "healthcare", "generator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fhir.resources>=7.1.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
    "httpx>=0.24",
    "faker>=18.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
kindling = "kindling.cli:main"
kindling-validate = "kindling.validate_cli:validate"

[tool.setuptools.packages.find]
where = ["."]
include = ["kindling*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true